+++ R/rate_functions.R
@@ -1048,7 +1048,7 @@ getFlux.MizerParams <- function(params, n = initialN(params),
     no_w <- length(params@w)
     
     g <- getEGrowth(params, n = n, n_pp = n_pp, n_other = n_other, t = t)
-    d <- params@diffusion
+    d <- params@ext_diffusion
     dw <- params@dw
     
     flux <- matrix(0, nrow = no_sp, ncol = no_w,